+++ db/migrate/20160819154017_add_completer_id_to_stickies.rb
@@ -1,0 +1,6 @@
+class AddCompleterIdToStickies < ActiveRecord::Migration[5.0]
+  def change
+    add_column :stickies, :completer_id, :integer
+    add_index :stickies, :completer_id
+  end
+end